Moved interfaces to types folder
