--- retrovirtualcockpit-web/src/configs/index.ts
@@ -1,1 +1,0 @@
-export * from './cockpit.interfaces'
