+++ retrovirtualcockpit-web/src/types/cockpit.interfaces.ts
