+++ retrovirtualcockpit-web/src/types/message.ts
@@ -1,0 +1,6 @@
+export interface Message {
+  key: string
+  modifier: string
+  action: string
+  autoKeyUpDelay: Number
+}
